Adding charging commands, verified most of other commands, also commented out i2c low power mode stuff

+++ Apps/FlightSoftware/Watchdog/src/watchdog.c
@@ -217,7 +217,7 @@ void heaterControl(){
         loop_flags |= FLAG_TEMP_HIGH;
     }
 
-    uint16_t PWM_cycle = 0;
+    uint32_t PWM_cycle = 0;
     // P controller output
     // setpoint is slightly above desired temp because otherwise it will get stuck slightly below it
     if(loop_flags & FLAG_TEMP_LOW){
